SRM2Storage: make sure src_file is local for creating pipe

+++ Resources/Storage/SRM2Storage.py
@@ -705,18 +705,20 @@ class SRM2Storage( StorageBase ):
       nbstreams = 1
     gLogger.info( "SRM2Storage.__putFile: Using %d streams" % nbstreams )
     gLogger.info( "SRM2Storage.__putFile: Executing transfer of %s to %s" % ( src_url, dest_url ) )
-    
-    ## create pipe if file size is bigger than 32MB
-    pipeName = "%s-%s" % ( tempfile.mktemp(), os.path.basename( src_file ) )
-    try:
-      if sourceSize > 33554432: 
-        os.mkfifo( pipeName )
-        ret = shellCall( cmdSeq = "dd if=%s of=%s bs=%sM &" % ( src_file, pipeName, self.uploadBlockSizeMB ), timeout = 10 )
-        if ret["OK"]:
-          gLogger.debug("SRM2Storage.__putFile: Pipe %s created" % pipeName )
-          src_url = "file:%s" % pipeName
-    except OSError, error:
-      gLogger.error( "SRM2Storage.__putFile: Unable to create pipe: %s" % str(error) )
+
+    ## local file?
+    if src_url.startswith("file:"): 
+      ## create pipe if file size is bigger than 32MB
+      pipeName = "%s-%s" % ( tempfile.mktemp(), os.path.basename( src_file ) )
+      try:
+        if sourceSize > 33554432: 
+          os.mkfifo( pipeName )
+          ret = shellCall( cmdSeq = "dd if=%s of=%s bs=%sM &" % ( src_file, pipeName, self.uploadBlockSizeMB ), timeout = 10 )
+          if ret["OK"]:
+            gLogger.debug("SRM2Storage.__putFile: Pipe %s created" % pipeName )
+            src_url = "file:%s" % pipeName
+      except OSError, error:
+        gLogger.error( "SRM2Storage.__putFile: Unable to create pipe: %s" % str(error) )
 
     res = pythonCall( ( timeout + 10 ), self.__lcg_cp_wrapper, src_url, dest_url,
                       srctype, dsttype, nbstreams, timeout, src_spacetokendesc, dest_spacetokendesc )
